New SummaryTab component file to handle state
